+++ TESTS/run.lua
@@ -59,6 +59,7 @@ local specs = {
   "fences_spec.lua",
   "fence_hl_spec.lua",
   "fence_content_hl_spec.lua",
+  "fence_jump_spec.lua",
   "fence_export_spec.lua",
   "fence_actions_spec.lua",
   "debounce_manager_spec.lua",